@@ -1,6 +1,6 @@
 module open-cluster-management.io/cert-policy-controller
 
-go 1.18
+go 1.19
 
 require (
 	github.com/go-logr/zapr v1.2.3
